Fixed test framework
